flextok_ar/pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "flextok-ar"
version = "0.1.0"
requires-python = ">=3.10"
readme = {file = "README.md", content-type = "text/markdown"}
license = {file = "LICENSE"}
dependencies = [
    # Core dependencies
    "flextok @ git+https://github.com/apple/ml-flextok.git",
    "l3m @ git+https://github.com/apple/ml-l3m.git",
    
    # Note: torch and torchvision are not listed here to avoid conflicts
    # with custom builds (e.g., NVIDIA builds). They should be installed
    # separately or will be pulled in by flextok dependency.
    # Minimum required: torch>=2.5.0, torchvision>=0.20.0
    
    # Configuration and utilities
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "fvcore>=0.1.5",
    
    # Array and image processing
    "numpy>=1.24.0",
    "pillow>=10.0.0",
    "einops>=0.7.0",
    
    # Model loading and serialization
    "safetensors>=0.4.0",
    "huggingface_hub>=0.20.0",
    
    # Text encoder (for T2I)
    "transformers>=4.30.0",
    "sentencepiece>=0.1.99",
    "accelerate>=0.20.0",
]

[tool.setuptools.packages.find]
where = [".."]
include = ["flextok_ar*"]

